- Added a number of regression tests
- Made the time and memory limit of tools configurable

- Added PrintTool for lpspp, bespp and pbespp


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13441 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tests/regression/test.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 #~ Copyright 2015 Sjoerd Cranen.
 #~ Copyright 2015 Wieger Wesselink.
 #~ Distributed under the Boost Software License, Version 1.0.
@@ -19,43 +21,40 @@ class TestRunner(testrunner.TestRunner):
         abspath = self.abspath
         self.test_path = os.path.join(os.getcwd(), os.path.dirname(__file__))
         pbessolve_files = sorted(os.listdir(os.path.join(self.test_path, 'pbessolve')))
-        self.tests = [('regression_pbessolve_{}'.format(file[:-4]), run_yml_test, [ymlfile('pbessolve'), [abspath('pbessolve/{}'.format(file))]]) for file in pbessolve_files] + \
+        self.tests = [
+                       ('ticket_283' , run_yml_test,             [ ymlfile('mcrl22lps'),         [abspath('tickets/283/1.mcrl2')]                                       ] ),
+                       ('ticket_325' , run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/325/1.mcrl2')]                                       ] ),
+                       ('ticket_1090', run_yml_test,             [ ymlfile('ticket_1090'),       [abspath('tickets/1090/1.mcrl2'), abspath('tickets/1090/1.mcf')]       ] ),
+                       ('ticket_1093', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1093/1.mcrl2')]                                      ], { 'timeout': 300, 'memlimit': 500000000 }),
+                       ('ticket_1122', run_yml_test,             [ ymlfile('ticket_1122'),       [abspath('tickets/1122/1.mcrl2')]                                      ] ),
+                       ('ticket_1127', run_yml_test,             [ ymlfile('mcrl22lps'),         [abspath('tickets/1127/1.mcrl2')]                                      ] ),
+                       ('ticket_1143', run_yml_test,             [ ymlfile('pbesrewr-onepoint'), [abspath('tickets/1143/1.txt')]                                        ] ),
+                       ('ticket_1144', run_yml_test,             [ ymlfile('lpsbisim2pbes'),     [abspath('tickets/1144/test1.txt'), abspath('tickets/1144/test2.txt')] ] ),
+                       ('ticket_1167', run_countstates_test,     [ [mcrl2file('/examples/academic/abp/abp.mcrl2')], 74                                                  ] ),
+                       ('ticket_1234', run_yml_test,             [ ymlfile('lpsbinary'),         [mcrl2file('/examples/academic/cabp/cabp.mcrl2')]                      ] ),
+                       ('ticket_1241', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1241/1.mcrl2')]                                      ] ),
+                       ('ticket_1249', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1249/1.mcrl2')]                                      ] ),
+                       ('ticket_1315', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1315/1.mcrl2')]                                      ] ),
+                       ('ticket_1316', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1316/1.mcrl2')]                                      ] ),
+                       ('ticket_1317', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1317/1.mcrl2')]                                      ] ),
+                       ('ticket_1318', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1318/1.mcrl2')]                                      ] ),
+                       ('ticket_1319', run_yml_test,             [ ymlfile('alphabet'),          [abspath('tickets/1319/1.mcrl2')]                                      ] ),
+                       ('lpsconfcheck_1', run_lpsconfcheck_ctau_test, [ [mcrl2file('/examples/academic/cabp/cabp.mcrl2')], ['-t8', '-xT'], (0, 18)                      ] ),
+                       ('lpsconfcheck_2', run_lpsconfcheck_ctau_test, [ [mcrl2file('/examples/academic/trains/trains.mcrl2')], ['-t8', '-xT'], (0, 9)                   ] ),
+                       #('lpsconfcheck_3', run_lpsconfcheck_ctau_test, [ [mcrl2file('/examples/industrial/chatbox/chatbox.mcrl2')], ['-t8', '-xT', '-u'], (40, 72)  ], { 'timeout': 300, 'memlimit': 500000000 } ),
+                     ] + \
                      [
-                       ('regression_ticket_283' , run_yml_test,         [ ymlfile('mcrl22lps'),         [abspath('tickets/283/1.mcrl2')]                                       ] ),
-                       ('regression_ticket_325' , run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/325/1.mcrl2')]                                       ] ),
-                       ('regression_ticket_1127', run_yml_test,         [ ymlfile('mcrl22lps'),         [abspath('tickets/1127/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1143', run_yml_test,         [ ymlfile('pbesrewr-onepoint'), [abspath('tickets/1143/1.txt')]                                        ] ),
-                       ('regression_ticket_1144', run_yml_test,         [ ymlfile('lpsbisim2pbes'),     [abspath('tickets/1144/test1.txt'), abspath('tickets/1144/test2.txt')] ] ),
-                       ('regression_ticket_1167', run_countstates_test, [ [mcrl2file('/examples/academic/abp/abp.mcrl2')], 74                                                  ] ),
-                       ('regression_ticket_1234', run_yml_test,         [ ymlfile('lpsbinary'),         [mcrl2file('/examples/academic/cabp/cabp.mcrl2')]                      ] ),
-                       ('regression_ticket_1241', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1241/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1249', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1249/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1315', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1315/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1316', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1316/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1317', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1317/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1318', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1318/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1319', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1319/1.mcrl2')]                                      ] ),
-                       ('regression_ticket_1321', run_yml_test,         [ ymlfile('alphabet'),          [abspath('tickets/1321/1.mcrl2')]                                      ] ),
+                       ('regression_pbessolve_{}'.format(file[:-4]), run_yml_test, [ymlfile('pbessolve'), [abspath('pbessolve/{}'.format(file))]]) for file in pbessolve_files
                      ]
-                     #(run_yml_test, 'lpsconfcheck_triangular_cabp', 'lpsconfcheck_triangular', [self._source_path + '/examples/academic/cabp/cabp.mcrl2']),
-                     #(run_yml_test, 'lpsconfcheck_triangular_trains', 'lpsconfcheck_triangular', [self._source_path + '/examples/academic/trains/trains.mcrl2']),
-                     #(run_yml_test, 'lpsconfcheck_triangular_no_sums_trains', 'lpsconfcheck_triangular_no_sums', [self._source_path + '/examples/academic/trains/trains.mcrl2']),
-                     # + \
-                     #generate_test('lpsconfcheck_triangular_cabp', 'run_lpsconfcheck_test', [self._source_path + '/examples/academic/cabp/cabp.mcrl2'], ['-xT'], (0, 10)) + \
-                     #[]
-                     # TODO: increase timeout value
-                     #('lpsconfcheck_triangular_no_sums_chatbox', 'lpsconfcheck_triangular_no_sums', [self._source_path + '/examples/industrial/chatbox/chatbox.mcrl2']),
-                     #('ticket_1247', 'alphabet', ['tickets/1247/1.mcrl2']),
-                     #('ticket_352',  'alphabet', ['tickets/352/1.mcrl2']),
-                     #('ticket_397',  'mcrl22lps', ['tickets/397/1.mcrl2']),
-                     #('ticket_1090', 'lps2pbes', ['tickets/1090/form.mcf', 'tickets/1090/spec.mcrl2']),
-                     # TODO: increase timeout value
-                     #('ticket_1093', 'alphabet', ['tickets/1093/1.mcrl2']),
-                     #('ticket_1122', 'txt2lps', ['tickets/1122/1.mcrl2']),
 
         self.settings = {'toolpath': self._tool_path,
                          'verbose': self._args.verbose,
-                         'cleanup_files': True}
+                         'cleanup_files': not self._args.keep_files}
+
+    def main(self):
+        if self._args.print_names:
+            self.print_names()
+        super(TestRunner, self).main()
 
     def ymlfile(self, name):
         return '{}/tests/specifications/{}.yml'.format(self._source_path, name)
@@ -68,17 +67,25 @@ class TestRunner(testrunner.TestRunner):
 
     def _get_commandline_parser(self):
         parser = super(TestRunner, self)._get_commandline_parser()
-        parser.add_argument('--verbose', dest='verbose', action='store_true', help='Display additional progress messages.')
+        parser.add_argument('-v', '--verbose', dest='verbose', action='store_true', help='Display additional progress messages.')
+        parser.add_argument('-k', '--keep-files', dest='keep_files', action='store_true', help='Keep the files produced by the test')
+        parser.add_argument('-p', '--print-names', dest='print_names', action='store_true', help='Print the names and the numbers of the tests')
         return parser
 
     def names(self):
         for test in self.tests:
             yield test[0]
 
+    # displays names and numbers of the tests
+    def print_names(self):
+        for i, name in enumerate(self.names()):
+            print '{} {}'.format(i, name)
+
     def run(self, testnum):
         if testnum < len(self.tests):
             test = self.tests[testnum]
             name, testfunc, arguments = test[:3]
+            name = 'regression_{}'.format(name)
 
             # Create a settings dict based on self.settings and, optionally, the test-specific
             # settings
@@ -101,10 +108,11 @@ class TestRunner(testrunner.TestRunner):
             # might have created in the test directory, so we only have to remove an empty
             # directory here.
             os.chdir('..')
-            os.rmdir(name)
+            if not self._args.keep_files:
+                os.rmdir(name)
         else:
             raise RuntimeError('Invalid test number')
 
 if __name__ == "__main__":
     sys.exit(TestRunner().main())
-    #TestRunner().run(5)
+    #TestRunner().run(7)
